+++ App_Function_Libraries/Gradio_UI/Gradio_Shared.py
@@ -14,7 +14,7 @@ import gradio as gr
 # Local Imports
 from App_Function_Libraries.DB.DB_Manager import list_prompts, db, search_and_display, fetch_prompt_details
 from App_Function_Libraries.DB.SQLite_DB import DatabaseError
-from App_Function_Libraries.Utils import format_transcription
+from App_Function_Libraries.Utils.Utils import format_transcription
 #
 ##############################################################################################################
 #
