+++ App_Function_Libraries/Gradio_UI/Transcript_comparison.py
@@ -10,7 +10,7 @@ import gradio as gr
 
 from App_Function_Libraries.DB.DB_Manager import get_transcripts
 from App_Function_Libraries.Gradio_UI.Gradio_Shared import browse_items
-from App_Function_Libraries.Utils import format_transcription
+from App_Function_Libraries.Utils.Utils import format_transcription
 
 
 #
