+++ App_Function_Libraries/Gradio_UI/Utilities.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import gradio as gr
 import yt_dlp
 
-from App_Function_Libraries.Utils import sanitize_filename, downloaded_files
+from App_Function_Libraries.Utils.Utils import sanitize_filename, downloaded_files
 
 
 def create_utilities_yt_video_tab():
