+++ App_Function_Libraries/LLM_API_Calls_Local.py
@@ -21,7 +21,7 @@ from typing import Union
 ####################
 # Import necessary libraries
 # Import Local
-from Utils import *
+from App_Function_Libraries.Utils.Utils import *
 #
 #######################################################################################################################
 # Function Definitions
