+++ App_Function_Libraries/Local_File_Processing_Lib.py
@@ -23,7 +23,7 @@
 from App_Function_Libraries.Audio_Transcription_Lib import convert_to_wav
 from App_Function_Libraries.Video_DL_Ingestion_Lib import *
 from App_Function_Libraries.Video_DL_Ingestion_Lib import get_youtube
-from App_Function_Libraries.Utils import normalize_title, create_download_directory
+from App_Function_Libraries.Utils.Utils import normalize_title, create_download_directory
 
 #######################################################################################################################
 # Function Definitions
